% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth-data.R
\docType{data}
\name{ContaminationSummary}
\alias{ContaminationSummary}
\title{Results from the Ocean Genomics Contamintaion Analysis}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 40 rows and 4 columns.
}
\source{
\href{https://txomeai.oceangenomics.com/}{Ocean Genomics}
}
\usage{
ContaminationSummary
}
\description{
Results from the Ocean Genomics Contamintaion Analysis
}
\examples{
data(ContaminationSummary)
head(ContaminationSummary)
}
\keyword{datasets}
