% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth-data.R
\docType{data}
\name{DifferentialGeneExpression}
\alias{DifferentialGeneExpression}
\title{Results from the Ocean Genomics Differential Gene Expression analysis.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 38043 rows and 8 columns.
}
\source{
\href{https://txomeai.oceangenomics.com/}{Ocean Genomics}
}
\usage{
DifferentialGeneExpression
}
\description{
Results from the Ocean Genomics Differential Gene Expression analysis.
}
\examples{
data(DifferentialGeneExpression)
head(DifferentialGeneExpression)
}
\keyword{datasets}
