% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth-data.R
\docType{data}
\name{HighVarianceGenes}
\alias{HighVarianceGenes}
\title{Results from the Ocean Genomics High Variance Gene analysis.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1 rows and 6 columns.
}
\source{
\href{https://txomeai.oceangenomics.com/}{Ocean Genomics}
}
\usage{
HighVarianceGenes
}
\description{
Results from the Ocean Genomics High Variance Gene analysis.
}
\examples{
data(HighVarianceGenes)
head(HighVarianceGenes)
}
\keyword{datasets}
