% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth-data.R
\docType{data}
\name{SpearmanCorrelation}
\alias{SpearmanCorrelation}
\title{Results from the Ocean Genomics Spearman Correlation analysis.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 10 rows and 12 columns.
}
\source{
\href{https://txomeai.oceangenomics.com/}{Ocean Genomics}
}
\usage{
SpearmanCorrelation
}
\description{
Results from the Ocean Genomics Spearman Correlation analysis.
}
\examples{
data(SpearmanCorrelation)
head(SpearmanCorrelation)
}
\keyword{datasets}
