% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth-data.R
\docType{data}
\name{synthetic_data_report}
\alias{synthetic_data_report}
\title{Synthetic RNA-Seq data run on the Ocean Genomics platform}
\format{
An object of class \code{list} of length 8.
}
\source{
\href{https://txomeai.oceangenomics.com/}{Ocean Genomics}
}
\usage{
synthetic_data_report
}
\description{
Synthetic RNA-seq data run against the Human Ensemble v101 database.
}
\examples{
data(synthetic_data_report)
head(synthetic_data_report$ls)
}
\keyword{datasets}
