% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth-data.R
\docType{data}
\name{TranscriptIntegrityStats}
\alias{TranscriptIntegrityStats}
\title{Results from the Ocean Genomics Transcript Integrity analysis.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 10 rows and 6 columns.
}
\source{
\href{https://txomeai.oceangenomics.com/}{Ocean Genomics}
}
\usage{
TranscriptIntegrityStats
}
\description{
Results from the Ocean Genomics Transcript Integrity analysis.
}
\examples{
data(TranscriptIntegrityStats)
head(TranscriptIntegrityStats)
}
\keyword{datasets}
