\name{txomeai-package}
\alias{txomeai-package}
\alias{txomeai}
\docType{package}
\title{
\packageTitle{txomeai}
}
\description{
\packageDescription{txomeai}
THis package is the API to access report data generated by the txomeai tool from Ocean Genomics.
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{txomeai}
\packageIndices{txomeai}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{txomeai}

Maintainer: \packageMaintainer{txomeai}
Joshua P. Katz support@oceangenomics.com
}
\references{
https://txomeai.oceangenomics.com/
}
\keyword{ package }
\examples{
report = txomeai_connect("https://txomeai.oceangenomics.com/api/pipeline-output/c444dfda-de51-4053-8cb7-881dd1b2734d/2021-10-25T185916/report/index.html", dir = ".")
report_ls = txomeai_ls(report)
report_sv = txomeai_get_all(report, "StructuralVariants")
report_samples = txomeai_get(report, report_ls[report_ls$name == "sampleName",])
}
