% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txomeai.R
\name{txomeai_connect}
\alias{txomeai_connect}
\title{Build a connection object to a RNA-seq analysis report}
\usage{
txomeai_connect(url)
}
\arguments{
\item{url}{The report URL to connect to}
}
\value{
The constructed connection object
 \item{url}{The URL used to access the API data.}
 \item{CAS}{The ID of the analysis.}
 \item{instance}{The ID of the report.}
 \item{dir}{The path to the query save location.}
 \item{data}{A table of all the report page results.}
 \item{meta}{A table with each row representing a data file.}
 \item{sample}{A table with each row a sample and each column a data file.}
 \item{ls}{A summary table of all the available data.}
}
\description{
Build a connection object to a RNA-seq analysis report
}
\examples{
domain = "https://txomeai.oceangenomics.com"
path = "api/pipeline-output/c444dfda-de51-4053-8cb7-881dd1b2734d/2021-10-25T185916/report/index.html"
report = txomeai_connect(paste(domain, path, sep="/"))
}
\references{
https://txomeai.oceangenomics.com/
}
