% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txomeai.R
\name{txomeai_get}
\alias{txomeai_get}
\title{Get data from the Txome.AI anaylsis report.}
\usage{
txomeai_get(txomeai, tableName, tableKey = NULL)
}
\arguments{
\item{txomeai}{The connection object returned from txomeai_connect.}

\item{tableName}{A name column value from the ls table.}

\item{tableKey}{An optional key value from the ls table.}
}
\value{
a data.table with all results
 \item{txomeai}{The connection object returned from txomeai_connect.}
 \item{tableName}{A name column value from the ls table.}
 \item{tableKey}{An optional key value from the ls table.}
}
\description{
Get data from the Txome.AI anaylsis report.
}
