% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txomeai.R
\name{txomeai_login}
\alias{txomeai_login}
\title{Used to login, only use this function if you have a connection object and authentication is failing}
\usage{
txomeai_login(txomeai)
}
\arguments{
\item{txomeai}{The connection object}
}
\value{
True if login was successful, false otherwise.
}
\description{
Used to login, only use this function if you have a connection object and authentication is failing
}
