% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txomeai.R
\name{txomeai_ls}
\alias{txomeai_ls}
\title{This function returns a table of all available data in the report}
\usage{
txomeai_ls(txomeai)
}
\arguments{
\item{txomeai}{The report connection object}
}
\value{
A data.table that each row represents a data query from the report.
}
\description{
Returns a table with columns key, name, and description.
Each row represents a data query available in the report.
}
\examples{
domain = "https://txomeai.oceangenomics.com"
path = "api/pipeline-output/c444dfda-de51-4053-8cb7-881dd1b2734d/2021-10-25T185916/report/index.html"
report = txomeai_connect(paste(domain, path, sep="/"))
report_ls = txomeai_ls(report)
head(report_ls)
}
