% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txomeai.R
\name{txomeai_display}
\alias{txomeai_display}
\title{Display a svg diagram.}
\usage{
txomeai_display(txomeai, row)
}
\arguments{
\item{txomeai}{The connection object.}

\item{row}{A row from the txomeai$assets table.}
}
\value{
The SVG path if download is successful, null otherwise.
}
\description{
An experimental function to display a report SVG in whatever way is available.
}
\examples{
dontrun
domain <- "https://txomeai.oceangenomics.com"
path <- "api/pipeline-output/c444dfda-de51-4053-8cb7-881dd1b2734d/2021-10-25T185916/report/index.html"
report <- txomeai_connect(paste(domain, path, sep="/"))
txomeai_display(report, report$assets[1,])
}
