% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txomeai.R
\name{txomeai_display}
\alias{txomeai_display}
\title{Display a svg diagram.}
\usage{
txomeai_display(txomeai, ls_row)
}
\arguments{
\item{txomeai}{The connection object that cookie has expired.}

\item{ls_row}{An asset row from the txomeai_ls table.}
}
\value{
The SVG path if download is successful, null otherwise.
}
\description{
This function will display a downloaded SVG in whatever way is available.
}
\examples{
domain = "https://txomeai.oceangenomics.com"
path = "api/pipeline-output/c444dfda-de51-4053-8cb7-881dd1b2734d/2021-10-25T185916/report/index.html"
report = txomeai_connect(paste(domain, path, sep="/"))
report_ls = txomeai_ls(report)
assets = report_ls[report_ls$key == "assets", ]
txomeai_display(report, assets[1,])
}
