% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txomeai.R
\name{txomeai_get}
\alias{txomeai_get}
\title{Get data from the Txome.AI anaylsis report.}
\usage{
txomeai_get(connection, tableName, tableKey = NULL)
}
\arguments{
\item{connection}{The connection object returned from txomeai_connect.}

\item{tableName}{A name column value from the ls table (see txomeai_connect).}

\item{tableKey}{An optional key value from the ls table (see txomeai_connect).}
}
\value{
A data.table with all results.
}
\description{
Get data from the Txome.AI anaylsis report.
}
\examples{
dontrun
domain <- "https://txomeai.oceangenomics.com"
path <- "api/pipeline-output/c444dfda-de51-4053-8cb7-881dd1b2734d/2021-10-25T185916/report/index.html"
report <- txomeai_connect(paste(domain, path, sep="/"))
report_fastp <- txomeai_get(report, "FastpJSON")
}
